Pick up and place the Solar Panel
FINALLY. You can now pick up the Solar Panel, press it from the inventory and then place it (basically replacing the placeholder block). But now you cannot place it inside objects above the ground etc. You can only place it on the ground.

How it works (the changes) in short:
Placement calls upon the SO (Scriptable Object) and the Prefab of it to make this happen.