Added Damage Effect & Some UI Panels
When the player gets damaged/starts losing health the screen flashes red to show you are taking damage, before it was hard to notice that unless you had your eyes glued to the health and stamina bars.

Also added some Panels to serve as backgrounds for the UI elements (Health, Stamina and Pollution) so that the text underneath and the overall bars are easy to see. Before, it was hard to read/see when the camera was looking at darker objects in the scene, the text would blend in etc. But that is NO MORE :)