You can now win/complete the game!
I fixed the placement of the SolarPanel.
Placing the SolarPanel toggles on 2 buttons. "Done" and "Cancel". Upon pressing "Done" the pollution level decreases by 10 and you can no longer pick up/interact with the Solar Panel again and of course the buttons toggle off again.
If you press "Cancel" the buttons toggle off and you can pick up the SolarPanel again and place it again then you are faced with the 2 buttons again...

You can now interact with the Keyboards found at Wind Turbines Control Room and Nuclear Plant Control Room.

Upon hovering your mouse over one of the 2 keyboards a text saying "Press E to Turn Off" if you are at the one in the Nuclear Plant, and says "Press E to Turn ON" if you are at the one next to the Wind Turbines, appears and disappears when you hover your mouse away. And of course, pressing E does exactly that ^ and decreases the Pollution Level by 10 at each location.

Visual Changes:
When Pollution decreases, a green flash happens to show clearly that something happened, in this case, the pollution bar showed pollution decreasing. (Similar to the Health Damage overlay that happens upon getting damaged but in green).

TLDR; You can now win the game just lower the Pollution 80>50 and then you will go to/load the WinScene.